requirements.txt
# SelFlow - Self-Evolving AI Operating System Layer
# Core Dependencies

# Machine Learning
torch>=2.0.0
numpy>=1.24.0
transformers>=4.30.0
accelerate>=0.20.0
peft>=0.4.0

# GUI (Optional - for system tray)
PyQt6>=6.5.0

# macOS Integration
pyobjc-framework-Cocoa>=9.0
pyobjc-framework-Quartz>=9.0
pyobjc-framework-ApplicationServices>=9.0

# System Monitoring
psutil>=5.9.0
watchdog>=3.0.0

# Computer Vision & OCR
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=9.5.0

# Audio Processing (for future voice interface)
pyaudio>=0.2.11
speechrecognition>=3.10.0

# Data Processing
pandas>=2.0.0
scipy>=1.10.0

# Async & Networking
aiohttp>=3.8.0
aiofiles>=23.0.0
redis>=4.5.0

# Security & Privacy
cryptography>=41.0.0
keyring>=24.0.0

# Configuration & Utilities
PyYAML>=6.0
click>=8.1.0
tqdm>=4.65.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Logging & Monitoring
structlog>=23.1.0

# Core ML and AI
bitsandbytes>=0.42.0
datasets>=2.16.0

# Audio Processing
openai-whisper>=20231117
pvporcupine>=3.0.0  # Wake word detection
librosa>=0.10.1
soundfile>=0.12.1
webrtcvad>=2.0.10  # Voice activity detection

# Text-to-Speech
pyttsx3>=2.90
coqui-tts>=0.20.0

# System Control & Automation (macOS)
pyautogui>=0.9.54    # GUI automation
pynput>=1.7.6        # Keyboard/mouse control
applescript>=1.0.0   # Native macOS app control

# Screen Intelligence & Vision
pytesseract>=0.3.10  # OCR for screen reading
pillow>=10.1.0
easyocr>=1.7.0       # Alternative OCR
screeninfo>=0.8.1    # Screen detection

# UI Automation & Accessibility
atomacos>=3.0.0      # macOS accessibility API
selenium>=4.15.0     # Web browser automation
appscript>=1.2.0     # AppleScript from Python

# File System & Data Management
pathlib              # Built-in Python
shutil               # Built-in Python
zipfile              # Built-in Python

# Network & Web
requests>=2.31.0
beautifulsoup4>=4.12.0  # Web scraping
playwright>=1.40.0      # Modern web automation

# Database & Storage
sqlite3  # Built-in Python
sqlalchemy>=2.0.0
duckdb>=0.9.0  # Fast analytics database

# Configuration & Data Processing
numpy>=1.24.0
scipy>=1.11.0

# GUI Framework
plyer>=2.1.0   # Notifications

# Natural Language Processing
spacy>=3.7.0
sentence-transformers>=2.2.0
nltk>=3.8.1

# Computer Vision for UI Understanding
torch-audio>=2.1.0
torchvision>=0.16.0

# Workflow & Task Management
celery>=5.3.0        # Background task processing
schedule>=1.2.0      # Task scheduling
apscheduler>=3.10.0  # Advanced scheduling

# Security & Privacy
keyring>=24.3.0      # Secure credential storage

# Development & Testing
pytest>=7.4.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0

# Packaging & Distribution (macOS)
py2app>=0.28.0
dmgbuild>=1.6.0
pyinstaller>=6.0.0

# Performance & Monitoring
memory-profiler>=0.61.0
line-profiler>=4.1.0
objgraph>=3.6.0

# Optional: Advanced Features
faiss-cpu>=1.7.4     # Vector similarity search
chromadb>=0.4.0      # Vector database for memory
langchain>=0.1.0     # LLM orchestration framework 